move BannerShoesItem to folder molecules
