+++ components/molecules/BannerShoesItem.tsx
